feat(torghut): expose revenue repair digest
